Commas were being left out of JSON string fixed.
There was a conditional missing in the logic for generating the JSON
string for groups, causing commas to be ommited if there were fields
following groups listed. Fixed by checking to see if there are any
fields to be printed, and if there are none, delete the default
trailing comma.