Decided to do away with conditionals.
Decided conditionals are better implemented by users of the callback
mechanism instead of the framework, due to limitations of C++03 (I can't
make the functors generic enough-- I was being forced to specify a
number of parameters for the comparison function, or offload these to a
constructor, which frankly killed readability). As a result, we need to
have an example demonstrating how to make conditionals with the normal
callback system.

This commit removes all code that dealt with conditionals, but it keeps
the callback generalization. It also clarifies what the types of
functions the ouroboros_server registration expects.