REST callbacks now working.
There is a hack in place to work around the issue that callbacks are
currently limited to a single parameter of type var_type*. The server
keeps a static pointer to a server instance that gets passed to the
callback for sending REST callbacks. This means that running a second
server will probably result in undefined behavior. This is a limitation
of the current callback system.