Memory cleanups and improvements to interface
Changed things around so that we no longer have memory leaks. The
generator now generates a second function which is responsible for
freeing up all the memory allocated by build_tree.

In addition, group gained two new methods, that return a vector of the
fields and groups immediately within it.